+++ include/JSystem/JAudio2/JAUStreamAramMgr.h
@@ -64,7 +64,7 @@ public:
             if (!this->field_0x4.test(i)) {
                 continue;
             }
-            if ((u32)this->mHeaps[i].getBase() != param_0) {
+            if ((uintptr_t)this->mHeaps[i].getBase() != param_0) {
                 continue;
             }
             this->field_0x4.reset(i);
