+++ src/JSystem/J2DGraph/J2DWindowEx.cpp
@@ -920,7 +920,7 @@ void J2DWindowEx::setAnimation(J2DAnmVtxColor* param_0) {
                 for (u16 j = 0; j < uVar3; j++) {
                     J3DAnmVtxColorIndexData* puVar1 = param_0->getAnmVtxColorIndexData(0, j);
                     u16* indexPointer = param_0->getVtxColorIndexPointer(0);
-                    u16* indexPointer2 = indexPointer + (u32)puVar1->mpData;
+                    u16* indexPointer2 = indexPointer + (uintptr_t)puVar1->mpData;
                     for (u16 k = 0; k < puVar1->mNum; k++) {
                         if (indexPointer2[k] == field_0x168[i]) {
                             mAnmVtxColor = param_0;
@@ -964,7 +964,7 @@ const J2DAnmTransform* J2DWindowEx::animationPane(J2DAnmTransform const* param_0
                 for (u16 j = 0; j < uVar3; j++) {
                     J3DAnmVtxColorIndexData* puVar1 = mAnmVtxColor->getAnmVtxColorIndexData(0, j);
                     u16* indexPointer = mAnmVtxColor->getVtxColorIndexPointer(0);
-                    u16* indexPointer2 = indexPointer + (u32)puVar1->mpData;
+                    u16* indexPointer2 = indexPointer + (uintptr_t)puVar1->mpData;
                     for (u16 k = 0; k < puVar1->mNum; k++) {
                         if (indexPointer2[k] == field_0x168[i]) {
                             mAnmVtxColor->getColor(0, j, local_38[i]);
